@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: '>=3.2.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,22 +29,19 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  go_router: ^4.2.7
-#  flutter_riverpod: ^1.0.4
-  flutter_hooks: ^0.18.5+1
-  hooks_riverpod: ^1.0.4
-  json_annotation: ^4.6.0
-  hive_flutter: ^1.1.0
-#  flutter_secure_storage: ^5.1.0
-  smooth_page_indicator: ^1.0.0+2
-  flutter_html: ^2.2.1
-  flutter_screenutil: ^5.5.3+2
-#  path_provider: ^2.0.11
-  logger: ^1.1.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.6
+  go_router: ^13.0.1
+  flutter_hooks: ^0.20.3
+  hooks_riverpod: ^2.4.9
+  json_annotation: ^4.8.1
+  hive_flutter: ^1.1.0
+  smooth_page_indicator: ^1.1.0
+  flutter_html: ^3.0.0-beta.2
+  flutter_screenutil: ^5.9.0
+  logger: ^2.0.2+1
 
 dev_dependencies:
   flutter_test:
@@ -55,9 +52,9 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
-  build_runner: ^2.2.0
-  hive_generator: ^1.1.3
+  flutter_lints: ^3.0.1
+  build_runner: ^2.4.6
+  hive_generator: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
