@@ -1,29 +1,25 @@
+import 'package:cutaway/database/table/notification_info.dart';
+import 'package:cutaway/database/table/promotion_info.dart';
+import 'package:cutaway/router/route_utils.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-
-import '../../../database/table/notification_info.dart';
-import '../../../database/table/promotion_info.dart';
-import '../../../router/route_utils.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 part 'notification_view_model.dart';
 
 class NotificationPage extends ConsumerWidget {
-  NotificationPage({Key? key}) : super(key: key);
+  NotificationPage({super.key});
 
-  final List<NotificationInfo> notificationList =
-      Hive.box<NotificationInfo>(tableNotificationInfo).values.toList();
+  final List<NotificationInfo> notificationList = Hive.box<NotificationInfo>(tableNotificationInfo).values.toList();
 
-  final List<PromotionInfo> promotionList =
-      Hive.box<PromotionInfo>(tablePromotionInfo).values.toList();
+  final List<PromotionInfo> promotionList = Hive.box<PromotionInfo>(tablePromotionInfo).values.toList();
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Colors.white,
         iconTheme: const IconThemeData(color: Colors.black),
         centerTitle: true,
         title: const Text(
@@ -41,22 +37,21 @@ class NotificationPage extends ConsumerWidget {
                 Expanded(
                   flex: 1,
                   child: Consumer(builder: (context, ref, _) {
-                    var page = ref.watch(_pagePosition.state).state;
+                    var page = ref.watch(_pagePosition.notifier).state;
                     var color = page == 0 ? Colors.yellow[100] : Colors.white;
 
                     return TextButton(
-                      onPressed: () => ref.read(_pagePosition.state).state = 0,
+                      onPressed: () => ref.read(_pagePosition.notifier).state = 0,
                       style: TextButton.styleFrom(
                           backgroundColor: color,
                           side: const BorderSide(color: Colors.grey),
-                          shape: const RoundedRectangleBorder(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(4)))),
+                          shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(4)))),
                       child: const Center(
-                          child: Text(
-                        '最新消息',
-                        style: TextStyle(color: Colors.black),
-                      )),
+                        child: Text(
+                          '最新消息',
+                          style: TextStyle(color: Colors.black),
+                        ),
+                      ),
                     );
                   }),
                 ),
@@ -74,12 +69,13 @@ class NotificationPage extends ConsumerWidget {
                       style: TextButton.styleFrom(
                           backgroundColor: color,
                           side: const BorderSide(color: Colors.grey),
-                          shape: const RoundedRectangleBorder(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(4)))),
+                          shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(4)))),
                       child: const Center(
-                          child: Text('卡編專區',
-                              style: TextStyle(color: Colors.black))),
+                        child: Text(
+                          '卡編專區',
+                          style: TextStyle(color: Colors.black),
+                        ),
+                      ),
                     );
                   }),
                 ),
@@ -122,30 +118,30 @@ class ItemNotification extends StatelessWidget {
       padding: const EdgeInsets.symmetric(vertical: 8),
       child: InkWell(
         onTap: () {
-          var router = GoRouter.of(context);
-          router.go('${router.location}${AppPage.notificationDetail.fullPath}',
-              extra: _info);
+          context.pushNamed(AppPage.notificationDetail.name, extra: _info);
         },
         child: Ink(
           height: 80.sp,
           decoration: BoxDecoration(
-              color: Colors.white,
-              border: Border.all(),
-              borderRadius: const BorderRadius.all(Radius.circular(4.0))),
+              color: Colors.white, border: Border.all(), borderRadius: const BorderRadius.all(Radius.circular(4.0))),
           child: Padding(
             padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text(_info.title,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(color: Colors.black)),
+                Text(
+                  _info.title,
+                  maxLines: 2,
+                  overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(color: Colors.black),
+                ),
                 Expanded(child: Container()),
                 Align(
                   alignment: Alignment.bottomRight,
-                  child: Text(_getDateString(_info.sentTime),
-                      style: const TextStyle(color: Colors.black)),
+                  child: Text(
+                    _getDateString(_info.sentTime),
+                    style: const TextStyle(color: Colors.black),
+                  ),
                 )
               ],
             ),
@@ -162,7 +158,7 @@ class ItemNotification extends StatelessWidget {
 }
 
 class ItemPromotion extends StatelessWidget {
-  const ItemPromotion(this._info, {Key? key}) : super(key: key);
+  const ItemPromotion(this._info, {super.key});
 
   final PromotionInfo _info;
 
@@ -175,27 +171,31 @@ class ItemPromotion extends StatelessWidget {
         child: Ink(
           height: 100.sp,
           decoration: BoxDecoration(
-              color: Colors.white,
-              border: Border.all(),
-              borderRadius: const BorderRadius.all(Radius.circular(4.0))),
+              color: Colors.white, border: Border.all(), borderRadius: const BorderRadius.all(Radius.circular(4.0))),
           child: Padding(
             padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 12),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text(_info.title,
-                    maxLines: 1,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(color: Colors.black)),
-                Text(_info.description,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(color: Colors.black)),
+                Text(
+                  _info.title,
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(color: Colors.black),
+                ),
+                Text(
+                  _info.description,
+                  maxLines: 2,
+                  overflow: TextOverflow.ellipsis,
+                  style: const TextStyle(color: Colors.black),
+                ),
                 Expanded(child: Container()),
                 Align(
                   alignment: Alignment.bottomRight,
-                  child: Text(_getDateString(_info.sentTime),
-                      style: const TextStyle(color: Colors.black)),
+                  child: Text(
+                    _getDateString(_info.sentTime),
+                    style: const TextStyle(color: Colors.black),
+                  ),
                 )
               ],
             ),
