@@ -1,3 +1,3 @@
 part of 'home_page.dart';
 
-var _pagePosition = StateProvider.autoDispose<int>((ref) => 0);
+final _pagePosition = StateProvider.autoDispose<int>((ref) => 0);
