@@ -1,7 +1,7 @@
 part of 'login_page.dart';
 
-var _isAccountNotEmpty = StateProvider.autoDispose((ref) => false);
+final _isAccountNotEmpty = StateProvider.autoDispose((ref) => false);
 
-var _isPasswordNotEmpty = StateProvider.autoDispose((ref) => false);
+final _isPasswordNotEmpty = StateProvider.autoDispose((ref) => false);
 
-var _isPasswordVisible = StateProvider.autoDispose((ref) => false);
+final _isPasswordVisible = StateProvider.autoDispose((ref) => false);
